moved timeline aggregate directives to their own page
This is a branchable admin change to the content of this side, which I am
slightly uncomfortable with, but we noticed that this was using excessive
amounts of CPU.

It seems that the timeline page takes several minutes to build. Probably
due to the size of the page with everything inlined into it, as well
as needing to build the rss and atom feeds.

That's not a problem on its own; new items are relatively rarely added
to the page. However, the aggregate directives display the time of last aggregation,
and so that causes a rebuild of this page after every aggregation.

Branchable runs aggregations quite frequently -- every 5 minutes.
We've optimised several things to make aggregation speedy, but this frequency
means that as soon as it's done building the timeline page for one
aggregation, it's typically time for the next one.

I hope this change is not a bad surprise. Please contact admins@branchable.com
if you have any questions.