expire old aggregated tweets
We noticed that this page was taking a long time to build. The reason is that several thousand aggregated tweets have built up, and when the page builds, it has to pick the newest 55 to show. Since only 55 are show, I've taken the liberty of expiring all but the newest 100.